+++ ChangeLog.md
@@ -76,6 +76,9 @@ creates a more intuitive interface for new TurboVNC users.  A new command-line
 option (`-noautokill`), or the existing `$autokill` turbovncserver.conf
 variable, can be used to disable the feature.
 
+12. The TurboVNC Viewer's built-in SSH client can now read SSH private keys
+from ssh-agent.
+
 
 2.2.1
 =====